Given that actions are so annoying doing some changes
Not having separate option to trigger a single JOB manually
is so incredibly annoying and stupid.
People have their own ways of working and people
want to trigger jobs manually on their own.
Not only the whole stupid workflow/pipeline.
Let people do what they want and implement an
option to trigger jobs separately.
If your options with environments or other
are so good, then do tell them why.
Tell that there is this and that research,
these and those reasons why we think
that whatever else is better than triggering
jobs separately. Then people will be able to choose.
But let them choose.
Thus implement that option to trigger jobs separately.

There is already enough frustration in the world
and not having option to trigger jobs separately
is not a frustration that we need or that you couldn't fix.