Feature/tf workflows (#1)
* Trying different TF workflow

* Syntax fix

* version for action

* Comment out old action and try without base64

* So GitHub actions creates a failing workflow, if there is empty file

* Output test var

* Add environment, so the variables can be found

* Base64 for variables that will be files with quotes inside. Stupid actions

* init with plain terraform

* Some debug

* Lets see if the job works with parsed tf plan

* So I guess don't add : in string in actions?

* Manual apply

* Syntax - outputs

* Outputs again

* Debug

* Debug

* Gosh, forgot | to make run multiline

* Trying the updated output approach

* Bump version

* debug

* toJSON

* Keep debugging, because actions are so wonky

* jank

* trim newlines, because actions are so janky

* Given that actions are so annoying doing some changes

Not having separate option to trigger a single JOB manually
is so incredibly annoying and stupid.
People have their own ways of working and people
want to trigger jobs manually on their own.
Not only the whole stupid workflow/pipeline.
Let people do what they want and implement an
option to trigger jobs separately.
If your options with environments or other
are so good, then do tell them why.
Tell that there is this and that research,
these and those reasons why we think
that whatever else is better than triggering
jobs separately. Then people will be able to choose.
But let them choose.
Thus implement that option to trigger jobs separately.

There is already enough frustration in the world
and not having option to trigger jobs separately
is not a frustration that we need or that you couldn't fix.